refactor(use cases): rename all useCases directories to start with lower case and conform change
