+++ src/modules/car/useCases/listAvailableCars/ListAvailableCarsUseCase.spec.ts
