+++ src/modules/car/useCases/listAvailableCars/validateListAvailableCarsDTO.ts
