+++ src/modules/car/useCases/registerCar/RegisterCarDTO.ts
