+++ src/modules/car/useCases/registerCar/RegisterCarUseCase.spec.ts
