+++ src/modules/car/useCases/removeCar/RemoveCarController.ts
