+++ src/modules/car/useCases/removeCar/RemoveCarUseCase.ts
