+++ src/modules/rental/pipes/rentCar.pipe.ts
@@ -2,7 +2,7 @@ import { AppError } from '@shared/errors/AppError';
 import { extractValidationErrors } from '@shared/utils/formatValidationError';
 import { plainToClass } from 'class-transformer';
 import { validate } from 'class-validator';
-import { RentCarDTO } from '../useCases/RentCar/RentCarDTO';
+import { RentCarDTO } from '../useCases/rentCar/RentCarDTO';
 
 export const rentCarPipe = async (data: RentCarDTO): Promise<RentCarDTO> => {
   const rentCarDTO = plainToClass(RentCarDTO, data);
